Combined the two solution, worked on readability
Added intro, results, worked on the overall structure of the code. Combined the two codes with "Times". 
Worked locally on RStudio. Structure of the github repository still unclear. Melanie and I can not like or commit or accept changes of our own work.. 
Applied Google Style Guide for readability of code.