This week's homework about the Mean Boulder Size
The most likeli mean boulder size was computed in this homework. 
I used a for loop to calculate the likelihoods (3.5 to 6) in a stepsize of 0.1. Next I formed the product of the weighted priors and the likelihoods and normalized the term to the sum of the likelihoods. 

I am not quite sure whether I used the right normalizing approach, but my outcome fits the anticipated value of ~5.2 as a mean boulder size.
Let me know, if anyone has a better normalizing approach!